Re-introduce precise CRC checking for all span reads
I wanted to have it back at least partially because checking the
pages' CRC does not prevent us from having situations such as retrieving
correct data but for the wrong block. It'll be impossible to have this
checking in general but I figured it would be good to have it if possible.